tools: Improve proto annotation documentation
